requirements.txt
aiohttp==3.8.1
asyncpg==0.26.0
python-dotenv==0.21.0
